requirements.txt

streamlit==1.35.0
torch==2.3.0
transformers==4.41.2
sentence-transformers==2.7.0
qdrant-client==1.9.1


PyMuPDF==1.24.2
python-docx==1.1.0


psutil==5.9.8
GPUtil==1.4.0


tqdm==4.66.4
scikit-learn==1.5.0
portalocker==2.8.2  # required by qdrant-client

# Only needed if using quantized models (e.g. Phi-1.5 in GGUF etc.)
# sentencepiece==0.1.99
# bitsandbytes==0.42.0


docx2txt==0.8
